# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /auth/register:
    post:
      tags:
      - user-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseUserResponse"
  /auth/oauth/login:
    post:
      tags:
      - user-controller
      operationId: authenticateWithProvider
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginWithProviderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseUserResponse"
  /auth/login:
    post:
      tags:
      - user-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseUserResponse"
  /health:
    get:
      tags:
      - loady-application
      operationId: healthCheck
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseString"
  /auth/me:
    get:
      tags:
      - user-controller
      operationId: getCurrentUser
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseUserResponse"
  /auth/deleteAll:
    get:
      tags:
      - user-controller
      operationId: deleteAllUsers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseString"
  /:
    get:
      tags:
      - loady-application
      operationId: getMethodName
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    UserRequest:
      type: object
      properties:
        idToken:
          type: string
          maxLength: 4096
          minLength: 0
      required:
      - idToken
    ApiResponseUserResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/UserResponse"
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        status:
          type: integer
          format: int32
        metadata:
          type: object
          additionalProperties: {}
        errors:
          type: array
          items:
            type: string
    UserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - SUPER_ADMIN
    LoginWithProviderRequest:
      type: object
      properties:
        idToken:
          type: string
          maxLength: 4096
          minLength: 0
      required:
      - idToken
    ApiResponseString:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        status:
          type: integer
          format: int32
        metadata:
          type: object
          additionalProperties: {}
        errors:
          type: array
          items:
            type: string